Account deletion toevegoeved aan de admin pagina

+++ public/javascripts/deleteEvent.js
@@ -1,20 +1,24 @@
 document.querySelectorAll('.delete-btn').forEach(button => {
     button.addEventListener('click', function () {
         const eventId = this.getAttribute('data-event-id');
-        fetch(`/api/events/${eventId}`, {
-            method: 'DELETE'
-        })
-            .then(response => {
-                if (response.ok) {
-                    console.log('Event deleted successfully');
-                    window.location.reload();
-                } else {
-                    console.error('Failed to delete event');
-                }
+        const confirmation = confirm('Weet je zeker dat je dit evenement wilt verwijderen?');
+        
+        if (confirmation) {
+            fetch(`/api/events/${eventId}`, {
+                method: 'DELETE'
             })
-            .catch(error => {
-                console.error('Error:', error);
-            });
+                .then(response => {
+                    if (response.ok) {
+                        console.log('Event deleted successfully');
+                        window.location.reload();
+                    } else {
+                        console.error('Failed to delete event');
+                    }
+                })
+                .catch(error => {
+                    console.error('Error:', error);
+                });
+        }
     });
 });
 
@@ -24,5 +28,3 @@ const data = () => {
     const eventcode = document.getElementById('eventcode').innerText.trim()
     alert(`Deel je link\ ${url}?code=${eventcode}`);
 }
-
-
